solver function half completed!1

+++ src/elements/Union.java
@@ -7,7 +7,7 @@ public class Union {
     private ArrayList<Node> nodes;
     private boolean visited;
     private Node mainNode;
-    int I_p,I_n;
+    double I_p,I_n;
     /////////////////////////////////constructor///////////////////////////
     public Union(int number){
         this.number = number;
